Get git branches via filesystem
The idea behind this is that it's more reliable than calling a git
subprocess and parsing the output (what if people have aliased 'git
branch --list' to print some non-default format?). It also means we
aren't relying on (relatively large) libraries like git2 or gitoxide,
that we only use a very small fraction of (I'd like to keep the
executable < 1MB, if possible).