Fix path comparison in test on Windows
Using 'std::fs::canonicalize' on Windows seems to prepend the path with
'\\?\'; which enables Windows to deal with the path if it's >255
characters. That's fine, but tests that were comparing paths as strings
were failing because of it.

I considered just calling 'canonicalize' on the expected path in the
test, but did not want calling canonicalize on a path to be a
requirement of passing the test. I found the 'same_file' crate that
checks that two paths point to the same file on the system, which suits
the test's purpose.